requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-embed-video==1.3
django-gunicorn==0.1.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-paypal==1.0.0
django-youtube-api==0.0.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==20.0.4
heroku==0.1.4
httplib2==0.14.0
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==1.5
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
